+++ libs/i18n/src/index.ts
@@ -3,4 +3,5 @@ export * from './lib/i18n.module';
 export * from './lib/languages';
 export * from './lib/models';
 export * from './lib/pipes';
+export * from './lib/providers';
 export * from './lib/utils';